refactoring and cleanup
refactoring
fix of sonar lint recommendations
