+++ src/main/java/com/rj/ecommerce_backend/product/dtos/CategorySearchCriteria.java
@@ -1,7 +1,7 @@
 package com.rj.ecommerce_backend.product.dtos;
 
 import com.rj.ecommerce_backend.product.domain.Category;
-import com.rj.ecommerce_backend.product.mappers.CategorySpecifications;
+import com.rj.ecommerce_backend.product.search.CategorySpecifications;
 import org.springframework.data.jpa.domain.Specification;
 
 public record CategorySearchCriteria(
