+++ src/main/java/com/rj/ecommerce_backend/user/valueobject/PhoneNumber.java
@@ -4,6 +4,4 @@ import jakarta.persistence.Embeddable;
 
 @Embeddable
 public record PhoneNumber(String value) {
-//    private static final PhoneNumberUtil phoneUtil = PhoneNumberUtil.getInstance();
-    // TODO: check if there is any lib for phone validation
 }
